requirements.txt
python-telegram-bot==13.15
python-dotenv==1.1.1
gspread==6.2.1
flask==2.1.0  # Only if you use `keep_alive.py` with Flask
Werkzeug==2.0.3  # Required by Flask 2.1.0 to avoid 'url_quote' error
oauth2client==4.1.3
google-api-python-client==2.126.0